+++ __tests__/fixtures/empty/output.js
@@ -1,7 +1,0 @@
-import "@stylexjs/stylex";
-const {
-  className: s
-} = {};
-export {
-  s as className
-};