+++ __tests__/fixtures/path-alias/output.js
@@ -1,9 +1,0 @@
-import "@stylexjs/stylex";
-const {
-  className
-} = {
-  className: "x97kcnl"
-};
-export {
-  className
-};